refactored installer to make it more modular and easier to maintain

Resolved Conflicts:
	kagenti/installer/src/installer.py
