+++ kagenti/installer/app/checker.py
@@ -1,0 +1,109 @@
+# Assisted by watsonx Code Assistant
+# Copyright 2025 IBM Corp.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+import os
+import time
+
+import docker
+import typer
+from dotenv import load_dotenv
+from packaging.version import parse
+from rich.panel import Panel
+from rich.text import Text
+
+from . import config
+from .utils import console, get_command_version
+
+
+def check_dependencies():
+    """Checks if required command-line tools are installed and meet version requirements."""
+    console.print(
+        Panel(Text("1. Checking Dependencies", justify="center", style="bold yellow"))
+    )
+    all_ok = True
+    for tool, versions in config.REQ_VERSIONS.items():
+        with console.status(f"[cyan]Checking for {tool}..."):
+            time.sleep(0.5)
+            version = get_command_version(tool)
+            min_ver, max_ver = parse(versions["min"]), parse(versions["max"])
+            if version is None:
+                console.log(
+                    f"[bold red]✗ {tool}[/bold red] is not installed or not in PATH."
+                )
+                all_ok = False
+            elif not (min_ver <= version < max_ver):
+                console.log(
+                    f"[bold red]✗ {tool}[/bold red] version [bold yellow]{version}[/bold yellow] not in range ({min_ver} - {max_ver})."
+                )
+                all_ok = False
+            else:
+                console.log(
+                    f"[bold green]✓ {tool}[/bold green] version [bold cyan]{version}[/bold cyan] is compatible."
+                )
+    if not all_ok:
+        console.print(
+            "\n[bold red]Please install or update the required tools before proceeding.[/bold red]"
+        )
+        raise typer.Exit(1)
+    console.print("[bold green]All dependency checks passed.[/bold green]\n")
+
+
+def check_env_vars():
+    """Checks for the presence of required environment variables in the .env file."""
+    console.print(
+        Panel(
+            Text(
+                "2. Checking Environment Variables",
+                justify="center",
+                style="bold yellow",
+            )
+        )
+    )
+    with console.status("[cyan]Checking for .env file and variables..."):
+        time.sleep(0.5)
+        load_dotenv(dotenv_path=config.ENV_FILE)
+        required_vars = [
+            "GITHUB_USER",
+            "GITHUB_TOKEN",
+            "OPENAI_API_KEY",
+            "AGENT_NAMESPACES",
+        ]
+        missing = [v for v in required_vars if not os.getenv(v)]
+        if missing:
+            console.log(
+                f"[bold red]✗ Missing required environment variables: {', '.join(missing)}[/bold red]"
+            )
+            console.log(
+                "  Please ensure they are set in your .env file or environment."
+            )
+            raise typer.Exit(1)
+        console.log(
+            "[bold green]✓[/bold green] All required environment variables are set."
+        )
+    console.print()
+
+
+def kind_cluster_exists() -> bool:
+    """Checks if the Kind cluster is already running via the Docker daemon."""
+    try:
+        docker_client = docker.from_env()
+        return any(
+            config.CLUSTER_NAME in c.name for c in docker_client.containers.list()
+        )
+    except docker.errors.DockerException:
+        console.log(
+            "[bold red]✗ Docker is not running. Please start the Docker daemon.[/bold red]"
+        )
+        raise typer.Exit(1)
