+++ kagenti/installer/app/resources/gateway-nodeport.yaml
