+++ kagenti/installer/app/resources/gateway-waypoint.yaml
