+++ kagenti/installer/app/resources/http-gateway.yaml
