+++ kagenti/installer/app/resources/keycloak-namespace.yaml
