+++ kagenti/installer/app/resources/keycloak-route.yaml
