+++ kagenti/installer/app/resources/kiali-route.yaml
