+++ kagenti/installer/app/resources/otel-collector.yaml
