+++ kagenti/installer/app/resources/phoenix.yaml
