+++ kagenti/installer/app/resources/registry.yaml
