+++ kagenti/installer/pyproject.toml
@@ -18,11 +18,11 @@ dependencies = [
     "docker>=7.0.0,<8.0.0",
     "typer[all]>=0.16.0",
     "packaging>=23.0,<25.0",
-    "python-keycloak>=5.5.1",
+    "requests>=2.30.0,<3.0.0",
+    "python-keycloak>=5.5.1"
 ]
 
 [project.scripts]
-kagenti-installer = "src.installer:app"
+kagenti-installer = "app.cli:app"
 
-[tool.setuptools]
-packages = ["src"]
+[tool.setuptools.packages.find]
