+++ kagenti/ui/lib/a2a_utils.py
@@ -1,3 +1,4 @@
+# Assisted by watsonx Code Assistant
 # Copyright 2025 IBM Corp.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -12,7 +13,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-# Assisted by watsonx Code Assistant
 import streamlit as st
 import httpx
 import logging
